Function sp_consensus_beefy::check_commitment_signature

pub fn check_commitment_signature<Number, Id, MsgHash>(
    commitment: &Commitment<Number>,
    authority_id: &Id,
    signature: &<Id as RuntimeAppPublic>::Signature,
) -> bool
where
    Id: BeefyAuthorityId<MsgHash>,
    Number: Clone + Encode + PartialEq,
    MsgHash: Hash,

Check a commitment signature by encoding the commitment and verifying the provided signature using the expected authority id.