sp_consensus_beefy

Function check_double_voting_proof

pub fn check_double_voting_proof<Number, Id, MsgHash>(
    report: &DoubleVotingProof<Number, Id, <Id as RuntimeAppPublic>::Signature>,
) -> bool
where
    Id: BeefyAuthorityId<MsgHash> + PartialEq,
    Number: Clone + Encode + PartialEq,
    MsgHash: Hash,

Verifies the equivocation proof by making sure that both votes target different blocks and that its signatures are valid.