sp_consensus_beefy::mmr

Trait BeefyDataProvider

pub trait BeefyDataProvider<ExtraData> {
    // Required method
    fn extra_data() -> ExtraData;
}

A provider for extra data that gets added to the Mmr leaf

Required Methods

fn extra_data() -> ExtraData

Return a vector of bytes, ideally should be a merkle root hash

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.