Module sp_consensus_beefy::witness

Primitives for light, 2-phase interactive verification protocol.

Instead of submitting full list of signatures, it’s possible to submit first a witness form of SignedCommitment. This can later be verified by the client requesting only some (out of all) signatures for verification. This allows lowering the data and computation cost of verifying the signed commitment.

Structs

SignedCommitmentWitness

A light form of SignedCommitment.