Module sp_consensus_beefy::mmr

BEEFY + MMR utilities.

While BEEFY can be used completely independently as an additional consensus gadget, it is designed around a main use case of bridging standalone networks together. For that use case it’s common to use some aggregated data structure (like MMR) to be used in conjunction with BEEFY, to be able to efficiently prove any past blockchain data.

This module contains primitives used by Polkadot implementation of the BEEFY+MMR bridge, but we imagine they will be useful for other chains that either want to bridge with Polkadot or are completely standalone, but heavily inspired by Polkadot.

Structs

BeefyAuthoritySet

Details of a BEEFY authority set.

MmrLeaf

A standard leaf that gets added every block to the MMR constructed by Substrate’s pallet_mmr.

MmrLeafVersion

An MMR leaf versioning scheme.

MmrRootProvider

A crate::Payload provider where payload is Merkle Mountain Range root hash.

Traits

BeefyDataProvider

A provider for extra data that gets added to the Mmr leaf

Functions

find_mmr_root_digest

Extract the MMR root hash from a digest in the given header, if it exists.

Type Aliases

BeefyNextAuthoritySet

Details of the next BEEFY authority set.