Trait sp_consensus_beefy::test_utils::BeefySignerAuthority

pub trait BeefySignerAuthority<MsgHash: Hash>: AppPair {
    // Required method
    fn sign_with_hasher(&self, message: &[u8]) -> <Self as AppCrypto>::Signature;
}

Trait representing BEEFY specific generation and signing behavior of authority id

Accepts custom hashing fn for the message and custom convertor fn for the signer.

Required Methods

fn sign_with_hasher(&self, message: &[u8]) -> <Self as AppCrypto>::Signature

Generate and return signature for message using custom hashing MsgHash

Object Safety

This trait is not object safe.