Trait sp_consensus_beefy::AuthorityIdBound

pub trait AuthorityIdBound: Ord + AppPublic + Display + BeefyAuthorityId<BeefySignatureHasher, Signature = Self::BoundedSignature> {
    type BoundedSignature: Debug + Eq + PartialEq + Clone + TypeInfo + Codec + Send + Sync;
}

A trait bound which lists all traits which are required to be implemented by a BEEFY AuthorityId type in order to be able to be used in BEEFY Keystore

Required Associated Types

type BoundedSignature: Debug + Eq + PartialEq + Clone + TypeInfo + Codec + Send + Sync

Necessary bounds on the Signature associated with the AuthorityId

Object Safety

This trait is not object safe.