Trait sp_consensus_beefy::BeefyAuthorityId

pub trait BeefyAuthorityId<MsgHash: Hash>: RuntimeAppPublic {
    // Required method
    fn verify(
        &self,
        signature: &<Self as RuntimeAppPublic>::Signature,
        msg: &[u8],
    ) -> bool;
}

Trait representing BEEFY authority id, including custom signature verification.

Accepts custom hashing fn for the message and custom convertor fn for the signer.

Required Methods

fn verify( &self, signature: &<Self as RuntimeAppPublic>::Signature, msg: &[u8], ) -> bool

Verify a signature.

Return true if signature over msg is valid for this id.

Object Safety

This trait is not object safe.